requirements.txt
Flask
python-dotenv
GitPython
flask-login
werkzeug
requests
